+++ packages/third-party-dts-extractor/src/ThirdPartyExtractor.spec.ts
@@ -12,12 +12,17 @@ describe('ThirdPartyExtractor', () => {
     exclude: ['ignore-pkg', /ignore-pkg2-/, /ignore-pkg3/.toString()],
   });
 
-  it('should correctly infer pkg dir with types field in package.json', () => {
+  it("should correctly infer pkg's types dir with types/typings field in package.json", () => {
     const tsupDir = thirdPartyExtractor.inferPkgDir('tsup');
-    const pkgJson = fs.readJSONSync(`${tsupDir}/package.json`);
 
-    expect(pkgJson.name).toBe('tsup');
-    expect(Boolean(pkgJson.types || pkgJson.typings)).toEqual(true);
+    if (!tsupDir) {
+      throw new Error('tsup dir not found');
+    }
+
+    const dirContent = fs.readdirSync(tsupDir);
+    const dtsFiles = dirContent.filter((file) => file.endsWith('.d.ts'));
+
+    expect(dtsFiles.length).toBeGreaterThan(0);
   });
 
   it('should correctly infer pkg types dir without types field in package.json', () => {